Home_work_10
- added captcha to Registration form;
- added images to articles.
