+++ composer.json
@@ -12,7 +12,8 @@
         "intervention/imagecache": "^2.3",
         "laravel/framework": "5.4.*",
         "laravel/tinker": "~1.0",
-        "predis/predis": "^1.1"
+        "predis/predis": "^1.1",
+        "mews/captcha": "dev-master"
     },
     "require-dev": {
         "mockery/mockery": "0.9.*",
