+++ config/imagerules.php
@@ -1,0 +1,23 @@
+<?php
+
+return [
+    'maxSize' => 10 * 1024 * 1024,
+    'minSize' => 10 * 1024,
+    'allowedExt' => [
+        'jpeg',
+        'jpg',
+        'png',
+        'gif',
+        'bmp',
+        'tiff',
+        'pdf'
+    ],
+    'allowedMime' => [
+        'image/jpeg',
+        'image/png',
+        'image/gif',
+        'image/bmp',
+        'image/tiff',
+        'application/pdf'
+    ],
+];
